Update Utilities_Container.js
updates... fixed up makeNewSpotColor() for efficiency i think..

updated afc() argument from "crit" to "childType" because wtf does crit mean in that context? parent[criteria]? no. we want childType. Now that's more clear. Also removed the array of explicitly allowed parent and child types.. Not necessary. Just check to see whether parent[childType] exists and is valid.. If not, alert and exit. No need to predefine everything. it's a pain in the ass.

updated findChildByName() to look for an exact match (insensitive) instead of a match anywhere in the parent (this caused issues like getting swatch "C10" instead of swatch "C1" because c10 was found first and technically matched the pattern "C1"..

added afo() function which takes in an object and returns and array of the top level properties. currently intended only for plain json objects formatted like the builder/order data from netsuite that's used for build mockup and build prod file.