Update Utilities_Container.js
fixed my updated getVisibleBounds function to work with clipping masks that are made out of live text. previously this relied on finding the clipping path and using that path to get the bounds while ignoring everything else in the clip group (because anything else inside that group would be smaller than the clip path, or clipped by it). but when text is used as the clipping mask, this fails.